+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### Master
 
+* New methodology for unread indicator and marking conversation as read - matt
 * Fixes Inbox loading indicator stuck and conversation row collapsing - luc
 * Optimize the Home/ForYou view by re-enabling removal of clipped subviews - alloy
 * Brings back artist rails on the Home/ForYou view - alloy